Fixed some implicit narrowing conversion issues with trivial type changes.

+++ bigdata-core/bigdata/src/java/com/bigdata/btree/IndexSegmentPlan.java
@@ -197,7 +197,7 @@ public class IndexSegmentPlan {
          */
         long nchildren = nleaves;
 
-        int nnodes = 0;
+        long nnodes = 0;
         
         for (int h = height - 1; h >= 0; h--) {
